Only issue is the light/dark theme button doesn't change colors correctly, but it does change the theme, otherwise everything is functioning properly

+++ lib/dashboard.dart
@@ -15,18 +15,18 @@ class Dashboard extends ConsumerStatefulWidget {
 }
 
 class DashboardState extends ConsumerState<Dashboard> {
-  late int balance = ref.read(currentUserNotifierProvider).userBalance;
-  late int debt = ref.read(currentUserNotifierProvider).userDebt;
-  int _balanceChange = 0;
+  late double balance = ref.read(currentUserNotifierProvider).userBalance;
+  late double debt = ref.read(currentUserNotifierProvider).userDebt;
+  double _balanceChange = 0;
   TextEditingController _balanceChangeTEC = TextEditingController();
 
   void error() {}
 
-  int add(int a, int b) => (a + b);
+  double add(double a, double b) => (a + b);
 
-  int subtract(int a, int b) => (a - b);
+  double subtract(double a, double b) => (a - b);
 
-  int loan(int a, int b) => (a - b);
+  double loan(double a, double b) => (a - b);
 
   void computeDeposit() {
     // a = balance, b = balanceChange, c = debt
@@ -61,12 +61,12 @@ class DashboardState extends ConsumerState<Dashboard> {
   }
 
   void deposit() {
-    _balanceChange = int.tryParse(_balanceChangeTEC.text) ?? 0;
+    _balanceChange = double.tryParse(_balanceChangeTEC.text) ?? 0;
     _balanceChange < 0 ? error() : computeDeposit();
   }
 
   void withdraw() {
-    _balanceChange = int.tryParse(_balanceChangeTEC.text) ?? 0;
+    _balanceChange = double.tryParse(_balanceChangeTEC.text) ?? 0;
     _balanceChange < 0 ? error() : computeWithdraw();
   }
